Further improve code and reduce sketch size

+++ src/types.h
@@ -2,7 +2,7 @@
 
 struct particleSensorState_t
 {
-    unsigned long avgPM25;
-    unsigned long measurements[5] = {0, 0, 0, 0, 0};
-    int measurementIdx = 0;
+    uint16_t avgPM25;
+    uint16_t measurements[5] = {0, 0, 0, 0, 0};
+    uint8_t measurementIdx = 0;
 };
